requirements.txt
fastapi
uvicorn[standard]
pydantic
pytesseract
Pillow
pdf2image
motor
python-dotenv
